fastapi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib
bcrypt